playing with rust book ch 7
