+++ .gitignore
@@ -14,3 +14,4 @@ target/
 *.iml
 
 .vscode/
+.DS_Store
