+++ Cargo.toml
@@ -1,10 +1,69 @@
+[package]
+name = "pyrsia"
+authors = ["Mitali Bisht <mitalib@jfrog.com>", "Joeri Sykora <joeri@sertik.net>", "Elliott Frisch <elliottf@jfrog.com>", "Chris Mc <christopherm@jfrog.com>"]
+version = "0.1.0"
+edition = "2021"
+
+[[bin]]
+name = "pyrsia"
+path = "src/bin/cli/main.rs"
+
+[[bin]]
+name = "pyrsia_node"
+path = "src/bin/node/main.rs"
+
+[dependencies]
+clap = { version = "3.0.0-beta.5", features = ["yaml"] }
+anyhow = "1.0.51"
+bincode = "1.3.3"
+blake3 = "1.2.0"
+bytes = "1.1.0"
+ctor = "0.1.21"
+defaults = "0.2.0"
+easy-hasher = "2.2.1"
+env_logger = "0.9.0"
+filename = "0.1.1"
+futures = "0.3.17"
+hex = "0.4.3"
+hyper = { version = "0.14", features = ["full"] }
+hyper-tls = "0.5.0"
+itertools = "0.10.3"
+lazy_static = "1.4.0"
+#libp2p = "0.40.0"
+libp2p = {git = "https://github.com/decentnetwork/rust-libp2p.git", branch = "fix-compiling-v0.40", features=["tcp-tokio"]}
+log = "0.4.14"
+once_cell = "1.5"
+pretty_env_logger = "0.4.0"
+rand = "0.8.4"
+reqwest = { version = "0.11.6", features = ["blocking", "json"] }
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
+sha2 = { version = "0.9.8" }
+signed_struct = { path = "src/signed_struct" }
+stringreader = "0.1.1"
+strum = "0.22.0"
+strum_macros = "0.22.0"
+tempfile = "3.2.0"
+test-log = "0.2.8"
+tokio = { version = "1.14.0", features = [ "macros", "rt-multi-thread", "io-std" ] }
+unqlite = "1.5.0"
+uuid = { version = "0.8.2", features = [ "v4" ] }
+warp = { version = "0.3.1", default-features = false }
+base64 = "0.13.0"
+detached-jws = "0.2.1"
+iso8601 = "0.4.1"
+openssl = "0.10.38"
+serde_jcs = "0.1.0"
+syn = { version = "1.0.82", features = ["derive", "parsing", "extra-traits"]}
+time = { version = "0.3.5", features= ["std", "formatting", "parsing"] }
+
+[dev-dependencies]
+expectest = "0.10.0"
+
 [workspace]
 
 members = [
-    "pyrsia_node",
-    "pyrsia_cli",
-    "pyrsia_node/pyrsia_client_lib",
-    "pyrsia_node/pyrsia_client_lib/signed_struct"
+    "src/signed_struct"
 ]
 
 [profile.dev]
