--- pyrsia_cli/Cargo.toml
@@ -1,19 +1,0 @@
-[package]
-name = "pyrsia_cli"
-authors = ["Mitali Bisht <mitalib@jfrog.com>"]
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-clap = { version = "3.0.0-beta.5", features = ["yaml"] }
-anyhow = "1.0"
-reqwest = { version = "0.11", features = ["json"] }
-tokio = { version = "1", features = [ "full"] }
-futures = { version = "0.3.*"}
-
-[[bin]]
-name = "pyrsia"
-path = "src/main.rs"
-
-[dev-dependencies]
-expectest = "0.10.0"
