--- pyrsia_node/Cargo.toml
@@ -1,50 +1,0 @@
-[package]
-name = "pyrsia_node"
-authors = ["Joeri Sykora <joeri@sertik.net>", "Elliott Frisch <elliottf@jfrog.com>", "Chris Mc <christopherm@jfrog.com>"]
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-anyhow = "1.0.45"
-bincode = "1.3.3"
-blake3 = "1.2.0"
-bytes = "1.1.0"
-clap = "2.33.0"
-ctor = "0.1.21"
-defaults = "0.2.0"
-easy-hasher = "2.2.1"
-env_logger = "0.9.0"
-filename = "0.1.1"
-futures = "0.3.17"
-hex = "0.4.3"
-hyper = { version = "0.14", features = ["full"] }
-hyper-tls = "0.5.0"
-itertools = "0.10.3"
-lazy_static = "1.4.0"
-#libp2p = "0.40.0"
-libp2p = {git = "https://github.com/decentnetwork/rust-libp2p.git", branch = "fix-compiling-v0.40", features=["tcp-tokio"]}
-log = "0.4.14"
-once_cell = "1.5"
-pretty_env_logger = "0.4.0"
-pyrsia_client_lib = { path = "pyrsia_client_lib" }
-rand = "0.8.4"
-reqwest = { version = "0.11.6", features = ["blocking", "json"] }
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0"
-sha2 = { version = "0.9.8" }
-signed_struct = { path = "pyrsia_client_lib/signed_struct" }
-stringreader = "0.1.1"
-strum = "0.22.0"
-strum_macros = "0.22.0"
-tempfile = "3.2.0"
-test-log = "0.2.8"
-tokio = { version = "1", features = [ "macros", "rt-multi-thread", "io-std" ] }
-unqlite = "1.5.0"
-uuid = { version = "0.8.2", features = [ "v4" ] }
-warp = { version = "0.3.1", default-features = false }
-
-[dev-dependencies]
-env_logger = "*"
-test_helpers = { path = "test_helpers" }
