--- pyrsia_node/pyrsia_client_lib/.gitignore
@@ -1,2 +1,0 @@
-# Because this is a library, we don't want to commit the Cargo.lock
-Cargo.lock
