--- pyrsia_node/pyrsia_client_lib/Cargo.toml
@@ -1,23 +1,0 @@
-[package]
-name = "pyrsia_client_lib"
-version = "0.1.0"
-edition = "2021"
-
-[lib]
-
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-anyhow = "1.0.48"
-base64 = "0.13.0"
-detached-jws = "0.2.1"
-env_logger = "0.9.0"
-iso8601 = "0.4.1"
-log = "0.4.14"
-openssl = "0.10.38"
-serde = { version = "1.0", features = ["derive"] }
-serde_jcs = "0.1.0"
-serde_json = "1.0"
-syn = { version = "1.0.82", features = ["derive", "parsing", "extra-traits"]}
-time = { version = "0.3.5", features= ["std", "formatting", "parsing"] }
