--- pyrsia_node/resources/test/artifact_test.txt
@@ -1,1 +1,0 @@
-This is the pyrsia node manager testing data. Pushing and pulling artifacts from node. This is good testing data
