+++ readme.md
@@ -21,8 +21,8 @@ Before getting started, take a moment to review our [contributing guidelines](ht
 
 There are two components of this project
 
-- **[CLI](pyrsia_cli/)**: A basic interface which communicates with a node.
-- **[Node](pyrsia_node/)**: An instance of the Pyrsia daemon which can participate in the network with other nodes.
+- **[CLI](src/bin/cli/)**: A basic interface which communicates with a node.
+- **[Node](src/bin/node/)**: An instance of the Pyrsia daemon which can participate in the network with other nodes.
 
 ### Getting Started
