+++ src/bin/cli/readme.md
