+++ src/docker/v2/handlers/blobs.rs
