+++ src/docker/v2/handlers/manifests.rs
