+++ src/metadata_manager.rs
