+++ src/node_manager/model.rs
