+++ src/node_manager/model/artifact.rs
@@ -13,7 +13,7 @@
    See the License for the specific language governing permissions and
    limitations under the License.
 */
-extern crate pyrsia_client_lib;
+
 extern crate serde;
 extern crate serde_json;
