+++ src/node_manager/model/package_version.rs
@@ -13,12 +13,12 @@
    See the License for the specific language governing permissions and
    limitations under the License.
 */
-extern crate pyrsia_client_lib;
+
 extern crate serde_json;
 
 use crate::node_manager::model::artifact::Artifact;
 use crate::node_manager::model::package_type::PackageTypeName;
-use pyrsia_client_lib::signed::Signed;
+use crate::signed::Signed;
 use serde::{Deserialize, Serialize};
 use serde_json::{Map, Value};
 use signed_struct::signed_struct;
