+++ tests/string_manipulation/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "test_helpers"
+name = "string_manipulation"
 version = "0.1.0"
 edition = "2021"
