+++ tests/string_manipulation/src/lib.rs
